Add a ".xml" suffix to config files.
Ensure we only use ".xml" files for config - avoids backups etc.

Rearrange the beans a bit. Now have a bootstrap bean that doesn't do much but loads all the others.

Provide option to specify the location of the old style options. Use that instead of that on the class path.

Provide a stripped down version of those options for deployment to the server